+++ tgbotapi.extensions.behaviour_builder/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/triggers_handling/EventTriggers.kt
@@ -24,60 +24,70 @@ internal suspend inline fun <reified T : ChatEvent> BehaviourContext.onEvent(
     (it.asBaseSentMessageUpdate() ?.data ?.asChatEventMessage() ?.takeIf { it.chatEvent is T } as? ChatEventMessage<T>) ?.let(::listOfNotNull)
 }
 
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onChannelEvent(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<ChannelEvent>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<ChannelEvent>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<ChannelEvent>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onChatEvent(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<ChatEvent>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<ChatEvent>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<ChatEvent>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onVoiceChatEvent(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<VoiceChatEvent>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<VoiceChatEvent>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<VoiceChatEvent>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onVoiceChatStartedEvent(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<VoiceChatStarted>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<VoiceChatStarted>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<VoiceChatStarted>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onVoiceChatEndedEvent(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<VoiceChatEnded>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<VoiceChatEnded>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<VoiceChatEnded>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onVoiceChatParticipantsInvitedEvent(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<VoiceChatParticipantsInvited>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<VoiceChatParticipantsInvited>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<VoiceChatParticipantsInvited>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onMessageAutoDeleteTimerChangedEvent(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<MessageAutoDeleteTimerChanged>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<MessageAutoDeleteTimerChanged>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<MessageAutoDeleteTimerChanged>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onCommonEvent(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<CommonEvent>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<CommonEvent>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<CommonEvent>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onGroupEvent(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<GroupEvent>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<GroupEvent>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<GroupEvent>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onSupergroupEvent(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<SupergroupEvent>>? = null,
@@ -85,60 +95,70 @@ suspend fun BehaviourContext.onSupergroupEvent(
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<SupergroupEvent>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
 
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onChannelChatCreated(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<ChannelChatCreated>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<ChannelChatCreated>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<ChannelChatCreated>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onDeleteChatPhoto(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<DeleteChatPhoto>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<DeleteChatPhoto>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<DeleteChatPhoto>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onGroupChatCreated(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<GroupChatCreated>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<GroupChatCreated>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<GroupChatCreated>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onLeftChatMember(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<LeftChatMember>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<LeftChatMember>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<LeftChatMember>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onNewChatMembers(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<NewChatMembers>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<NewChatMembers>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<NewChatMembers>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onNewChatPhoto(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<NewChatPhoto>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<NewChatPhoto>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<NewChatPhoto>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onNewChatTitle(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<NewChatTitle>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<NewChatTitle>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<NewChatTitle>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onPinnedMessage(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<PinnedMessage>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<PinnedMessage>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<PinnedMessage>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onProximityAlertTriggered(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<ProximityAlertTriggered>>? = null,
     markerFactory: MarkerFactory<in ChatEventMessage<ProximityAlertTriggered>, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, ChatEventMessage<ProximityAlertTriggered>>
 ) = onEvent(additionalFilter, if (includeFilterByChatInBehaviourSubContext) MessageFilterByChat else null, markerFactory, scenarioReceiver)
+@Deprecated(OldAPITriggersDeprecationText)
 suspend fun BehaviourContext.onSupergroupChatCreated(
     includeFilterByChatInBehaviourSubContext: Boolean,
     additionalFilter: SimpleFilter<ChatEventMessage<SupergroupChatCreated>>? = null,