+++ CHANGELOG.md
@@ -16,6 +16,7 @@
     * All triggers have been changed to use two filters: filter for in subcontext data and filter for incoming data
     * New waiters for edited content
     * New extension `BehaviourContext#followLocation`
+    * Old API for triggers with the flags like `includeFilterByChatInBehaviourSubContext` have been deprecated
 
 ## 0.35.8